refact: extracts components from SubmissionFeaturePage to be more readable
