+++ app/src/features/submissions/page/features/SubmissionFeaturePage.tsx
@@ -7,17 +7,7 @@ import Chip from '@mui/material/Chip';
 import CircularProgress from '@mui/material/CircularProgress';
 import Container from '@mui/material/Container';
 import Link from '@mui/material/Link';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-import Paper from '@mui/material/Paper';
 import Stack from '@mui/material/Stack';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableRow from '@mui/material/TableRow';
 import Typography from '@mui/material/Typography';
 import { PageHeader } from 'components/header/PageHeader';
 import { useApi } from 'hooks/useApi';
@@ -25,24 +15,9 @@ import { useCartContext } from 'hooks/useContext';
 import useDataLoader from 'hooks/useDataLoader';
 import { useEffect, useMemo } from 'react';
 import { Link as RouterLink, useParams } from 'react-router-dom';
+import { SubmissionFeatureProperties } from './components/SubmissionFeatureProperties';
+import { SubmissionFeatureRelated } from './components/SubmissionFeatureRelated';
 
-/**
- * Formats a property value for display.
- * Handles primitives, arrays, and objects.
- */
-const formatPropertyValue = (value: unknown): string => {
-  if (value === null || value === undefined) {
-    return '';
-  }
-  if (typeof value === 'object') {
-    return JSON.stringify(value);
-  }
-  return String(value);
-};
-
-/**
- * Feature detail page displaying properties and related features.
- */
 export const SubmissionFeaturePage = () => {
   const biohubApi = useApi();
   const { features: cartFeatures, addToCart, removeFromCart } = useCartContext();
@@ -62,7 +37,6 @@ export const SubmissionFeaturePage = () => {
     () => cartFeatures.some((f) => f.submission_feature_id === Number(submissionFeatureId)),
     [cartFeatures, submissionFeatureId]
   );
-
 
   if (!featureDataLoader.data) {
     return (
@@ -113,7 +87,9 @@ export const SubmissionFeaturePage = () => {
             <Link component={RouterLink} to="/search" underline="hover" color="inherit">
               Search
             </Link>
-            <Typography color="text.secondary">{feature.submission_name}</Typography>
+            <Link component={RouterLink} to={`/search/${feature.submission_id}`} underline="hover" color="inherit">
+              {feature.submission_name}
+            </Link>
             <Typography color="text.primary">{feature.feature_type_display_name}</Typography>
           </Breadcrumbs>
         }
@@ -135,60 +111,8 @@ export const SubmissionFeaturePage = () => {
       />
       <Container maxWidth="xl">
         <Stack spacing={3} py={4}>
-          {/* Properties Section */}
-          <Paper elevation={0} sx={{ border: '1px solid', borderColor: 'divider' }}>
-            <Box p={3}>
-              <Typography variant="h2" component="h2" mb={2}>
-                Properties
-              </Typography>
-              <TableContainer>
-                <Table size="small">
-                  <TableBody>
-                    {Object.entries(feature.data)
-                      .filter(([key]) => key !== 'geometry')
-                      .map(([key, value]) => (
-                        <TableRow key={key}>
-                          <TableCell
-                            component="th"
-                            scope="row"
-                            sx={{ fontWeight: 700, textTransform: 'capitalize', width: '30%' }}>
-                            {key.replace(/_/g, ' ')}
-                          </TableCell>
-                          <TableCell>{formatPropertyValue(value)}</TableCell>
-                        </TableRow>
-                      ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-            </Box>
-          </Paper>
-
-          {/* Related Features Section */}
-          <Paper elevation={0} sx={{ border: '1px solid', borderColor: 'divider' }}>
-            <Box p={3}>
-              <Typography variant="h2" component="h2" mb={2}>
-                Related
-              </Typography>
-              {relatedFeatures.length > 0 ? (
-                <List disablePadding>
-                  {relatedFeatures.map((related) => (
-                    <ListItem key={related.submission_feature_id} disablePadding divider>
-                      <ListItemButton
-                        component={RouterLink}
-                        to={`/search/${feature.submission_id}/feature/${related.submission_feature_id}`}>
-                        <ListItemText
-                          primary={related.data?.name || related.feature_type_display_name}
-                          secondary={related.feature_type_display_name}
-                        />
-                      </ListItemButton>
-                    </ListItem>
-                  ))}
-                </List>
-              ) : (
-                <Typography color="text.secondary">No related features</Typography>
-              )}
-            </Box>
-          </Paper>
+          <SubmissionFeatureProperties data={feature.data} />
+          <SubmissionFeatureRelated submissionId={feature.submission_id} relatedFeatures={relatedFeatures} />
         </Stack>
       </Container>
     </>
